ifo_types: remove zero values we don't read anymore
Either it really should be zero and it's checked on reading via SkipZeroBuf().
Either the value is unknown and we just skip it via SkipBuf().

The field values are not actually set, so should not be used by anyone.
We keep the for reference of the size and structures they correspond to.